Extract the Kubernetes backend into rise-backend-kubernetes, under BSL
The Kubernetes backend was the only one still living inside the root crate:
~7,400 lines across resource_builder.rs, webhook.rs, crd.rs, pods.rs,
identity_refresh.rs and ip_validator.rs, with no crate boundary to carry a
license of its own. It now mirrors rise-backend-docker and
rise-backend-ecs — a crate on the rise-backend-core seam — and carries
`license = "BUSL-1.1"` with its own LICENSE copy.

Nothing outside the crate reaches into it. The Metacontroller webhook
handlers take the WebhookContext introduced earlier, and the crate hands back
a fully-stated Router, so AppState never crosses the boundary. The old module
paths keep resolving through re-export shims, the same pattern the Docker
backend already uses.

Five configuration types move with the code they configure —
KubernetesLogBackendSettings, BackendAddress, CustomDomainTlsMode,
HealthProbeConfig and NetworkPolicyConfig, the last of which embeds
k8s_openapi types directly. settings.rs re-exports them, so the
DeploymentControllerSettings::Kubernetes variant is unchanged and the
generated config schema is byte-identical. So is the RiseProject CRD YAML;
both drift checks pass.

Two tests used the root crate's LocalEncryptionProvider only to check that
secret env vars round-trip. rise-backend-core::test_helpers now offers a
reversible stand-in, which is what those tests were actually asking for and
keeps a crypto dependency out of the shared crate.

The whole-suite count is unchanged at 1,515: 73 of them now run in the new
crate, and the Kubernetes log tests moved with the backend they cover.

No permissive crate depends on a BSL crate — verified against every
rise-backend-* and support crate. rise-deploy does, as it already did for
ECS: the server image links both and is covered by BSL, while the released
CLI builds default features and links neither.
