+++ crates/rise-backend-kubernetes/Cargo.toml
@@ -1,0 +1,50 @@
+[package]
+name = "rise-backend-kubernetes"
+version = "0.1.0"
+edition.workspace = true
+# Business Source License 1.1, not the workspace default. See ../../LICENSING.md.
+license = "BUSL-1.1"
+repository.workspace = true
+description = "Kubernetes deployment backend for Rise (Metacontroller webhook, resource builder, RiseProject CRD)"
+publish = false
+
+[features]
+default = []
+# Enables `schemars::JsonSchema` derives on the controller settings types so
+# `rise-deploy` can include them in the generated backend-config schema.
+schema = ["dep:schemars", "k8s-openapi/schemars", "rise-backend-core/schema"]
+
+[dependencies]
+anyhow = { workspace = true }
+async-stream = "0.3"
+async-trait = { workspace = true }
+# Must stay on the same major as `rise-deploy`: the root crate nests this
+# crate's Router into its own, and two axum majors would not compose.
+axum = "0.8.7"
+base64 = { workspace = true }
+chrono = { workspace = true }
+futures = "0.3"
+k8s-openapi = { version = "0.27", features = ["v1_31"] }
+kube = { version = "3", features = ["client", "derive", "runtime"] }
+rise-backend-auth = { path = "../rise-backend-auth" }
+rise-backend-core = { path = "../rise-backend-core" }
+rise-deployment-spec = { workspace = true }
+rise-runtime-sync = { path = "../rise-runtime-sync" }
+rustls = { version = "0.23", features = ["ring"] }
+schemars = { workspace = true, optional = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+sha2 = { workspace = true }
+# `PgPool` only, for the identity-refresh leader lease. This crate opens no
+# connection and runs no query -- all persistence goes through `DeploymentStore`.
+sqlx = { workspace = true }
+tokio = { workspace = true }
+tokio-util = "0.7"
+regex = "1.11"
+tracing = { workspace = true }
+url = { workspace = true }
+urlencoding = "2.1"
+uuid = { workspace = true }
+
+[dev-dependencies]
+tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
