+++ src/server/deployment/mod.rs
@@ -1,21 +1,30 @@
 pub mod controller;
 #[cfg(feature = "backend")]
-pub mod crd;
+pub mod crd {
+    pub use rise_backend_kubernetes::crd::*;
+}
 pub mod handlers;
 #[cfg(feature = "backend")]
-pub mod identity_refresh;
+pub mod identity_refresh {
+    pub use rise_backend_kubernetes::identity_refresh::*;
+}
 #[cfg(feature = "backend")]
-pub mod ip_validator;
+pub mod ip_validator {
+    pub use rise_backend_kubernetes::ip_validator::*;
+}
 #[cfg(feature = "backend")]
 pub mod logs;
 pub mod models;
-pub mod pods;
 #[cfg(feature = "backend")]
 pub mod quantity;
 #[cfg(feature = "backend")]
-pub mod resource_builder;
+pub mod resource_builder {
+    pub use rise_backend_kubernetes::resource_builder::*;
+}
 pub mod routes;
 pub mod state_machine;
 pub mod utils;
 #[cfg(feature = "backend")]
-pub mod webhook;
+pub mod webhook {
+    pub use rise_backend_kubernetes::webhook::*;
+}
